backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.1
razorpay==2.0.0
python-multipart==0.0.9
pydantic[email]==2.7.1
httpx==0.27.0
openai
PyMuPDF==1.24.4
docx2txt==0.8
